Move UserInfo to model package.
