+++ sdk/core/src/main/java/com/klaviyo/coresdk/Klaviyo.kt
@@ -7,6 +7,7 @@ import com.klaviyo.coresdk.model.KlaviyoEventType
 import com.klaviyo.coresdk.model.KlaviyoProfileAttributeKey
 import com.klaviyo.coresdk.model.Profile
 import com.klaviyo.coresdk.model.SharedPreferencesDataStore
+import com.klaviyo.coresdk.model.UserInfo
 import com.klaviyo.coresdk.networking.HttpKlaviyoApiClient
 import com.klaviyo.coresdk.networking.KlaviyoApiClient
 import com.klaviyo.coresdk.networking.KlaviyoNetworkMonitor
